FIX: right sidebar styles; FIX: left sidebar button hidden bug
