+++ brad-chat/src/components/ThemeChange.js
@@ -1,14 +1,7 @@
 import React, { useState, useEffect } from 'react';
 
 function ThemeChangeButton() {
-  const [theme, setTheme] = useState('light'); // Default to light mode
-  const [leftSidebar, setLeftSidebar] = useState('hidden'); // Default to light mode
-
-  const toggleLeftSideBar = () => {
-    const newLeftSidebar = leftSidebar === 'hidden' ? 'show' : 'hidden'
-    setLeftSidebar(newLeftSidebar)
-    localStorage.setItem('theme', newLeftSidebar); // Store theme in local storage
-  }
+  const [theme, setTheme] = useState('dark'); // Default to light mode
 
   const toggleTheme = () => {
     const newTheme = theme === 'light' ? 'dark' : 'light';
@@ -22,25 +15,8 @@ function ThemeChangeButton() {
     if (storedTheme) {
       setTheme(storedTheme);
     }
-
-    const storedleftSidebar = localStorage.getItem('leftSidebar');
-    if (storedleftSidebar) {
-      setTheme(storedleftSidebar);
-    }
   }, []);
 
-  useEffect(() => {
-    // Get the root element
-    document.getElementById("leftSideBarToggle").addEventListener("click", function() {
-      var div = document.getElementById("leftSideBar");
-      if (leftSidebar == 'hidden') {
-        div.classList.remove("hidden"); // Show the div
-      } else {
-        div.classList.add("hidden"); // Hide the div if clicked again
-      }
-    });
-
-  }, [leftSidebar]);
 
   useEffect(() => {
     // Get the root element
@@ -65,9 +41,6 @@ function ThemeChangeButton() {
 
   return (
     <div class='settings'>
-      <button id='leftSideBarToggle' class='toggle-left-side-bar' onClick={toggleLeftSideBar}>
-        {leftSidebar === 'hidden' ? 'Show Sidebar' : 'Hide Sidebar'}
-      </button>
       <button class='theme' onClick={toggleTheme}>
         {theme === 'light' ? 'Dark Mode' : 'Light Mode'}
       </button>
